// src/app/shared/models/app.ts
export interface App {
    appId: number
    appCode: string
    appName: string
    description: string
    subModuleCode: string
    icon: string
    routerLink: string
    appType: string
    isPrimary: boolean
    sequence: any
    createdById: number
    createdBy: string
    createdAt: string
    modifiedById: any
    modifiedBy: any
    modifiedAt: any
    deactivated: boolean
  }
  